User Service - Login API end point implement
User Service - Login API end point has been implemented. It checks if the payload has the correct JSON keys, will handle response if payload is incorrect. If payload is correct then check is username and password is valid. If valid it signs a JWT token using the key in the authentication folder, then sends response with JWT token.